# openApi/swagger.yaml
openapi: 3.0.0
info:
  title: 'LaserLiga API'
  description: 'All API methods on the Laser liga portal.'
  version: '1.0'
servers:
  -
    url: 'https://laserliga.cz'
paths:
...
  /api/import:
    post:
      tags:
        - Games
        - Import
      summary: 'Import results from a game file.'
      description: 'Import results from a game file.'
      operationId: importGameFile
      parameters:
        -
          name: players
          in: query
          description: 'Player meta information'
          required: false
          schema:
            type: array
            items:
              additionalProperties:
                properties:
                  name: { description: 'Full UTF-8 name of a player', type: string }
                  user: { description: "User's unique code", type: string }
                type: object
        -
          name: group
          in: query
          description: 'Game group ID'
          required: false
          schema:
            type: integer
      requestBody:
        description: "Game file's content."
        required: true
        content:
          text/plain: {  }
      responses:
        '201':
          description: 'Game imported'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '500':
          description: 'Server error during save operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/players:
    get:
      tags:
        - Players
      summary: 'Find Users'
      description: 'This method returns users based on the provided search parameters.'
      operationId: find
      parameters:
        -
          name: search
          in: query
          description: 'Search parameter (name, code, email)'
          required: false
          schema:
            type: string
        -
          name: arena
          in: query
          description: 'Home arena filter'
          required: false
          schema:
            oneOf:
              -
                type: integer
              -
                enum:
                  - self
        -
          name: connectionType
          in: query
          description: 'Connected account type filter'
          required: false
          schema:
            type: string
        -
          name: identifier
          in: query
          description: 'Connected account identifier'
          required: false
          schema:
            type: string
        -
          name: codes
          in: query
          description: 'List of user codes'
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: 'List of players'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LigaPlayer'
    post:
      tags:
        - Players
      summary: 'Register a new player'
      description: 'Register a new player'
      operationId: registerPlayer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: 'Lay zerteg'
                  nullable: false
                email:
                  type: string
                  format: email
                  example: lay@zerteg.com
                  nullable: false
                password:
                  type: string
                  example: superstrongpassword123
                  nullable: false
              type: object
      responses:
        '201':
          description: 'Created a new player'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LigaPlayer'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 'Internal error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Game:
      description: 'Base class for game models'
      properties:
        resultsFile:
          type: string
          nullable: true
        modeName:
          type: string
        fileTime:
          type: string
          format: date-time
          nullable: true
        start:
          type: string
          format: date-time
          nullable: true
        importTime:
          type: string
          format: date-time
          nullable: true
        end:
          type: string
          format: date-time
          nullable: true
        timing:
          nullable: true
        code:
          type: string
        mode:
          oneOf:
            -
              $ref: '#/components/schemas/AbstractMode'
          nullable: true
          description: $mode
        gameType: {  }
        arena:
          oneOf:
            -
              $ref: '#/components/schemas/Arena'
          nullable: true
        music:
          oneOf:
            -
              $ref: '#/components/schemas/MusicMode'
          nullable: true
        group:
          description: ''
          nullable: true
        started:
          type: boolean
        finished:
          type: boolean
        visited:
          type: boolean
      type: object
    AbstractMode:
      description: 'Base class for all game mode models'
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        type: {  }
        loadName:
          type: string
          nullable: true
        systems:
          type: string
          nullable: true
        teams:
          type: string
        settings:
          $ref: '#/components/schemas/ModeSettings'
        rankable:
          type: boolean
      type: object
...
    ModeSettings:
      description: 'Data-model for all game mode settings'
      properties:
        public:
          type: boolean
        mines:
          type: boolean
        partWin:
          type: boolean
        partTeams:
          type: boolean
        partPlayers:
          type: boolean
        partHits:
          type: boolean
        partBest:
          type: boolean
        partBestDay:
          type: boolean
        playerScore:
          type: boolean
        playerShots:
          type: boolean
        playerMiss:
          type: boolean
        playerAccuracy:
          type: boolean
        playerMines:
          type: boolean
        playerPlayers:
          type: boolean
        playerPlayersTeams:
          type: boolean
        playerKd:
          type: boolean
        playerFavourites:
          type: boolean
        playerLives:
          type: boolean
        teamScore:
          type: boolean
        teamAccuracy:
          type: boolean
        teamShots:
          type: boolean
        teamHits:
          type: boolean
        teamZakladny:
          type: boolean
        bestScore:
          type: boolean
        bestHits:
          type: boolean
        bestDeaths:
          type: boolean
        bestAccuracy:
          type: boolean
        bestHitsOwn:
          type: boolean
        bestDeathsOwn:
          type: boolean
        bestShots:
          type: boolean
        bestMiss:
          type: boolean
        bestMines:
          type: boolean
      type: object
...
    Title:
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        rarity:
          $ref: '#/components/schemas/Rarity'
        unlocked:
          type: boolean
      type: object
    Address:
      properties:
        street:
          type: string
          example: 'OD Luna 3. patro, Velké náměstí 175'
          nullable: true
        city:
          type: string
          example: Písek
          nullable: true
        postCode:
          type: string
          example: '39701'
          nullable: true
        country:
          type: string
          example: Česko
          nullable: true
      type: object
    Arena:
      properties:
        name:
          type: string
          example: 'Laser arena Písek'
        gameCodePrefix:
          type: string
          example: g
        lat:
          type: number
          format: float
          example: 49.307678
          nullable: true
        lng:
          type: number
          format: float
          example: 14.147773
          nullable: true
        address:
          $ref: '#/components/schemas/Address'
        web:
          type: string
          example: 'https://laserarenapisek.cz'
          nullable: true
        contactEmail:
          type: string
          example: info@laserarenapisek.cz
          nullable: true
        contactPhone:
          type: string
          example: '+420 776 606 631'
          nullable: true
      type: object
    LigaPlayer:
      description: 'Same as the regular player, but with the addition of the arena and user parameters'
      type: object
      allOf:
        -
          $ref: '#/components/schemas/LigaPlayerBase'
        -
          properties:
            user: {  }
            arena:
              oneOf:
                -
                  $ref: '#/components/schemas/Arena'
              nullable: true
          type: object
    LigaPlayerBase:
      properties:
        stats:
          $ref: '#/components/schemas/PlayerStats'
        code:
          description: 'Unique code for each player - two players can have the same code if they are from different arenas.'
          type: string
        nickname:
          type: string
        email:
          type: string
        avatar:
          type: string
          nullable: true
        avatarStyle:
          type: string
          nullable: true
        avatarSeed:
          type: string
          nullable: true
        title:
          oneOf:
            -
              $ref: '#/components/schemas/Title'
          nullable: true
        birthday:
          type: string
          format: date-time
          nullable: true
      type: object
...
    PlayerStats:
      properties:
        gamesPlayed:
          type: integer
        arenasPlayed:
          type: integer
        rank:
          type: integer
        averageAccuracy:
          type: number
          format: float
        averagePosition:
          type: number
          format: float
        maxAccuracy:
          type: integer
        maxScore:
          type: integer
        maxSkill:
          type: integer
        shots:
          type: integer
        averageShots:
          type: number
          format: float
        averageShotsPerMinute:
          type: number
          format: float
        totalMinutes:
          type: integer
        kd:
          type: number
          format: float
        hits:
          type: integer
        deaths:
          type: integer
      type: object
...
    MusicMode:
      properties:
        name:
          type: string
        order:
          type: integer
        fileName:
          type: string
        arena:
          oneOf:
            -
              $ref: '#/components/schemas/Arena'
          nullable: true
        idLocal:
          type: integer
        previewStart:
          type: integer
      type: object
    Rarity:
      type: string
      enum:
        - common
        - uncommon
        - rare
        - epic
        - legendary
        - mythic
...
    ErrorResponse:
      properties:
        title:
          type: string
          example: 'Error title'
        type:
          $ref: '#/components/schemas/ErrorType'
        detail:
          type: string
          example: 'Error description'
          nullable: true
        exception:
          properties:
            message:
              type: string
              example: 'Some exception description'
            code:
              type: integer
              example: 123
            trace:
              type: array
              items:
                type: object
              example:
                -
                  file: index.php
                  line: 1
                  function: abc
                  args: ['Argument value']
          type: object
          nullable: true
        values:
          type: object
          example:
            key1: value1
            key2: value2
          nullable: true
      type: object
...
    ErrorType:
      type: string
      enum:
        - validation_error
        - database_error
        - internal_error
        - resource_not_found_error
        - resource_access_error
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  -
    bearerAuth: []
...